Untangle test for http port
