--- tests/integration/test_prometheus_protocols/configs/prometheus_http_handlers_prefixed.xml
@@ -1,61 +1,0 @@
-<clickhouse>
-    <http_handlers>
-        <rule>
-            <url>/prometheus/api/v1/write</url>
-            <handler>
-                <type>prometheus_write</type>
-                <table>default.prometheus_http</table>
-            </handler>
-        </rule>
-        <rule>
-            <url>/prometheus/api/v1/read</url>
-            <handler>
-                <type>prometheus_read</type>
-                <table>default.prometheus_http</table>
-            </handler>
-        </rule>
-        <rule>
-            <url>/prometheus/api/v1/query_range</url>
-            <handler>
-                <type>prometheus_query</type>
-                <table>default.prometheus_http</table>
-            </handler>
-        </rule>
-        <rule>
-            <url>/prometheus/api/v1/query</url>
-            <handler>
-                <type>prometheus_query</type>
-                <table>default.prometheus_http</table>
-            </handler>
-        </rule>
-        <rule>
-            <url>/prometheus/api/v1/label/job/values</url>
-            <handler>
-                <type>prometheus_query</type>
-                <table>default.prometheus_http</table>
-            </handler>
-        </rule>
-        <rule>
-            <url>/api/v1/write</url>
-            <handler>
-                <type>prometheus_write</type>
-                <table>default.prometheus_http</table>
-            </handler>
-        </rule>
-        <rule>
-            <url>/api/v1/read</url>
-            <handler>
-                <type>prometheus_read</type>
-                <table>default.prometheus_http</table>
-            </handler>
-        </rule>
-        <rule>
-            <url>/api/v1/query</url>
-            <handler>
-                <type>prometheus_query</type>
-                <table>default.prometheus_http</table>
-            </handler>
-        </rule>
-        <defaults/>
-    </http_handlers>
-</clickhouse>
